feat: update Aviator function registration and introduce new context retrieval functions
