+++ src/main/java/org/phong/zenflow/workflow/subdomain/evaluator/functions/AviatorFunction.java
@@ -6,12 +6,12 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Marker annotation for Aviator functions that return a boolean value.
+ * Marker annotation for Aviator functions.
  * <p>
  * Classes annotated with this will be auto-registered by
  * {@link AviatorFunctionRegistry}.
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
-public @interface AviatorBooleanFunction {
+public @interface AviatorFunction {
 }
